refactor: rename menu to ham-menu
re #245
