+++ layouts/after-login/script.js
@@ -1,4 +1,4 @@
-import HamMenu from '~/components/menu/index.vue'
+import HamMenu from '~/components/ham-menu/index.vue'
 
 export default {
   components: {
